feat: fork executive process for monitoring and metrics reporting
Investigate idea of forking "executive" process for monitoring the
invocation of flox subcommands throughout their lifetime, and performing
monitoring and metrics submissions asynchronously so as to avoid
slowing down the straight-line performance of flox itself. This also
makes it possible to measure flox command durations and opens up
the possibility of gathering more detailed results.