feat: added support for fish and tcsh shells
Following the flox activate overhaul, it was fairly straightforward
to add support for the fish and tcsh shells, although I was surprised
to find the logic was spread across about a dozen files. I made a small
attempt to refactor and consolidate the shell-related assets into a
single `activate.d` package, but there are doubtless further improvements
to be made here.