feat: added support for fish and tcsh shells
Following the flox activate overhaul, it was fairly straightforward
to add support for the fish and tcsh shells, although I was surprised
to find the logic was spread across about a dozen files.

There are doubtless further improvements to be made here by way of
refactoring and consolidating shell-related assets into a single package.