fix: escape backticks as well when quoting cmd arguments
The perils of creating our own quoting function is that we're destined
to repeat lessons of the past, but of course the advantage is that we have
more precise control of the outcome. This patch adds logic to similarly
escape the backtick (`) character. (I rather suspect our next step will
be to escape the single quote ('), followed by conditional logic to avoid
escaping when within them.)