fix: FLOX_TURBO activation command mode
The [undocumented] FLOX_TURBO mode exists to launch a command without
invoking an intervening subshell or sourcing any of the manifest [profile]
scripts, and was not working as intended. This change makes it possible
to successfully invoke:

  % FLOX_TURBO=1 /path/to/activate bash <file>

In general the activate script should act like a shell and be able to
invoke commands with either of "activate <executable> <args>" or
"activate -c 'command args'" if wanting to source those commands with
a shell.