feat: use getopt, consistent -c handling for all cmd modes
PR1823 highlighted the reasoning behind our decision to invoke
commands following the `--` delimiter, and further established
this as exactly equivalent to the `-c` argument handling provided
by all shells. This patch implements this same handling natively
within the `activate` script, making it possible to invoke commands
using either of the `--` or `-c` notations in any of the `--turbo`,
`--noprofile`, or normal "cmd" modes.