feat: run builder.pl on user side, outside of sandbox
This update moves the invocation of builder.pl to the user side, prior
to invoking the derivation. The builder in the derivation then becomes
a shell script that simply copies the rendered files into place.

This also makes it possible to include a correct rendering of requisites.txt,
but at a cost. This currently bumps the env build time by more than 25%, so
I'll remove it in the next commit.